+++ src/Commands/Utilities/RecycleBinUtility.cs
@@ -1,9 +1,7 @@
 using Microsoft.SharePoint.Client;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
-using System.Text;
 
 namespace PnP.PowerShell.Commands.Utilities
 {
@@ -39,11 +37,11 @@ namespace PnP.PowerShell.Commands.Utilities
                 {
                     iterationRowLimit = 5000;
                 }
-                
+
                 items = ctx.Site.GetRecycleBinItems(pagingInfo, iterationRowLimit, false, RecycleBinOrderBy.DefaultOrderBy, recycleBinStage);
                 ctx.Load(items);
                 ctx.ExecuteQueryRetry();
-                recycleBinItems.AddRange(items.ToList());                                
+                recycleBinItems.AddRange(items.ToList());
 
                 // Paging magic (if needed)
                 // Based on this work our good friends at Portiva did ❤
@@ -60,6 +58,56 @@ namespace PnP.PowerShell.Commands.Utilities
             while (items?.Count == 5000); // if items had 5000 items, there might be more since that's the page size we're using
 
             return recycleBinItems;
+        }
+
+        internal static void RestoreOrClearRecycleBinItems(ClientContext ctx, int? rowLimit = null, RecycleBinItemState recycleBinItemState = RecycleBinItemState.None, bool restore = true)
+        {
+            string pagingInfo = null;
+            RecycleBinItemCollection items;
+
+            do
+            {
+                // We don't actually know what the List View Threshold for the Recycle Bin is, so we'll use the safe number (5000) and implement paging.
+                int iterationRowLimit;
+                if (rowLimit.HasValue && rowLimit.Value >= 5000)
+                {
+                    // Subtract this page's count from the rowLimit (we don't want duplicates or go out of bounds)
+                    if (rowLimit.HasValue) rowLimit -= 5000;
+
+                    iterationRowLimit = 5000;
+                }
+                else if (rowLimit.HasValue && rowLimit.Value > 0 && rowLimit.Value < 5000)
+                {
+                    iterationRowLimit = rowLimit.Value;
+                }
+                else
+                {
+                    iterationRowLimit = 5000;
+                }
+
+                items = ctx.Site.GetRecycleBinItems(pagingInfo, iterationRowLimit, false, RecycleBinOrderBy.DefaultOrderBy, recycleBinItemState);
+                ctx.Load(items);
+                ctx.ExecuteQueryRetry();
+
+                if (items.Count > 0)
+                {
+                    var nextId = items.Last().Id;
+                    var nextTitle = WebUtility.UrlEncode(items.Last().Title);
+                    pagingInfo = $"id={nextId}&title={nextTitle}";
+
+                    if (restore)
+                    {
+                        items.RestoreAll();
+                        ctx.ExecuteQueryRetry();
+                    }
+                    else
+                    {
+                        items.DeleteAll();
+                        ctx.ExecuteQueryRetry();
+                    }
+                }
+            }
+            while (items?.Count == 5000);
         }
     }
 }